Fixed away/home type mismatch bug

+++ src/Espn.Fantasy.Football.Domain/Model/Response/Matchup.cs
@@ -5,16 +5,16 @@ namespace Espn.Fantasy.Football.Domain.Model.Response
     public class Matchup
     {
         public int MatchupTypeId { get; set; }
-        public List<int> AwayTeamScores { get; set; }
+        public List<double> AwayTeamScores { get; set; }
         public Team AwayTeam { get; set; }
-        public int AwayTeamAdjustment { get; set; }
+        public double AwayTeamAdjustment { get; set; }
         public int AwayTeamId { get; set; }
         public bool IsBye { get; set; }
         public int HomeTeamId { get; set; }
-        public int HomeTeamAdjustment { get; set; }
-        public List<int> HomeTeamScores { get; set; }
-        public int HomeTeamBonus { get; set; }
-        public int Outcome { get; set; }
+        public double HomeTeamAdjustment { get; set; }
+        public List<double> HomeTeamScores { get; set; }
+        public double HomeTeamBonus { get; set; }
+        public double Outcome { get; set; }
         public Team HomeTeam { get; set; }
     }
 }
